+++ packages/cli/src/utils.ts
@@ -3,22 +3,18 @@ import type { PerformanceMeasure } from "node:perf_hooks";
 import type { PackageJson } from "type-fest";
 import { exec } from "node:child_process";
 import { createRequire } from "node:module";
+import { resolve } from "node:path";
 
 const require = createRequire(import.meta.url);
 export const packageJson = require("../package.json") as PackageJson;
 if (packageJson.bin === undefined) {
 	throw new Error("Script name is not defined.");
 }
-export const mainCommand = Object.keys(packageJson.bin)[0];
-export const testCommand = `${import.meta.dirname}/tests/node_modules/.bin/${mainCommand}`;
+export const testCommand = `node ${resolve(`${import.meta.dirname}/../dist/index.js`)}`;
 if (!packageJson.version) {
 	throw new Error("Package has no version number set.");
 }
 export const packageVersion = packageJson.version;
-
-export function hasHelpFlag(args: string[]) {
-	return !!args.find((arg) => arg === "--help" || arg === "-h");
-}
 
 export async function execute(
 	command: string,
@@ -40,117 +36,6 @@ export async function execute(
 			resolve({ error, stdout, stderr });
 		});
 	});
-}
-
-export function interpretFlag(
-	args: string[],
-	flag: string,
-	type: "boolean",
-):
-	| {
-			isPresent: false;
-			rawValue: null;
-			value: false;
-	  }
-	| {
-			isPresent: true;
-			rawValue: string | null;
-			value: boolean;
-	  };
-export function interpretFlag(
-	args: string[],
-	flag: string,
-	type: "list",
-):
-	| {
-			isPresent: false;
-			rawValue: null;
-			value: null;
-	  }
-	| {
-			isPresent: true;
-			rawValue: string | null;
-			value: string[];
-	  };
-export function interpretFlag(
-	args: string[],
-	flag: string,
-	type?: "string",
-):
-	| {
-			isPresent: false;
-			rawValue: null;
-			value: null;
-	  }
-	| {
-			isPresent: true;
-			rawValue: string | null;
-			value: string;
-	  };
-export function interpretFlag(
-	args: string[],
-	flag: string,
-	type: "string" | "boolean" | "list" = "string",
-):
-	| {
-			isPresent: false;
-			rawValue: null;
-			value: null;
-	  }
-	| {
-			isPresent: true;
-			rawValue: string | null;
-			value: string[];
-	  }
-	| {
-			isPresent: boolean;
-			rawValue: string | null;
-			value: boolean;
-	  }
-	| {
-			isPresent: true;
-			rawValue: string | null;
-			value: string | null;
-	  } {
-	const flagValue = args.find((arg) => arg.startsWith(flag));
-	if (flagValue === undefined) {
-		// Flag not found
-		return {
-			isPresent: false,
-			rawValue: null,
-			value: type === "boolean" ? false : null,
-		};
-	}
-	let rawValue: string | null = null;
-
-	if (flagValue.includes("=")) {
-		rawValue = flagValue.replace(`${flag}=`, "").trim();
-	} else {
-		const flagIndex = args.findIndex((arg) => arg === flag);
-		const potentialFlagValue = args[flagIndex + 1];
-		if (potentialFlagValue && !potentialFlagValue.startsWith("--")) {
-			rawValue = potentialFlagValue;
-		}
-	}
-	if (type === "boolean") {
-		return {
-			isPresent: true,
-			rawValue,
-			value: rawValue ? rawValue.toLowerCase() === "true" : true,
-		};
-	}
-	if (type === "list") {
-		return {
-			isPresent: true,
-			rawValue,
-			value: rawValue?.split(",") ?? [],
-		};
-	}
-	return {
-		isPresent: true,
-		rawValue,
-		value: rawValue,
-	};
 }
 
 export function toCamelCase(text: string) {