+++ packages/coding-standards/stylelint/index.mjs
@@ -1,3 +1,4 @@
-import { default as classicConfig } from "./classic.mjs";
-import { default as themeJsonConfig } from "./theme-json.mjs";
-export default { classicConfig, themeJsonConfig };
+import { default as classic } from "./classic.mjs";
+import { default as themeJson } from "./theme-json.mjs";
+import { default as tailwind } from "./tailwind.mjs";
+export default { classic, themeJson, tailwind };