+++ packages/coding-standards/stylelint/tailwind.mjs
@@ -1,0 +1,38 @@
+export default {
+	overrides: [
+		{
+			files: ["*.css", "**/*.css"],
+			rules: {
+				'at-rule-no-unknown': [
+				true,
+				{
+					ignoreAtRules: [
+						'tailwind',
+						'apply',
+						'variants',
+						'responsive',
+						'screen',
+					],
+				},
+				]
+			},
+		},
+		{
+			files: ["*.scss", "**/*.scss"],
+			rules: {
+				'scss/at-rule-no-unknown': [
+				true,
+				{
+					ignoreAtRules: [
+						'tailwind',
+						'apply',
+						'variants',
+						'responsive',
+						'screen',
+					],
+				},
+				]
+			},
+		},
+	]
+};